StegBot/requirements.txt
mysql-connector-python~=9.4.0
cryptography~=45.0.5
python-dotenv~=1.1.1
pillow~=11.3.0
piexif~=1.1.3
discord~=2.3.2
PyNaCl~=1.5.0